# use io.Copy for perf