for perf.

benchmarks:
`go test -timeout 3m -race -cover -v -run=XXXX -bench=BenchmarkPullDockerImage ./...`

// old 78163 ns/op, 81714 ns/op, 51060 ns/op

// new 29023 ns/op, 41531 ns/op, 30907 ns/op